fix more warnings

+++ src/lib/jp2/t1/Tier1.cpp
@@ -74,7 +74,7 @@ bool Tier1::encodeCodeblocks(grk_tcp *tcp,
 						block->cblk_sty = tccp->cblk_sty;
 						block->qmfbid = tccp->qmfbid;
 						block->resno = resno;
-						block->inv_step = band->inv_step;
+						block->inv_step = (int32_t)band->inv_step;
 						block->inv_step_ht = 1.0f/band->stepsize;
 						block->stepsize = band->stepsize;
 						block->x = (uint32_t)x;
